Use guint consistently for node ids
2004-06-24  Marco Pesenti Gritti  <marco@gnome.org>

        * lib/ephy-node.c: (ephy_node_new_with_id), (ephy_node_get_id),
        (child_changed), (write_parent):
        * lib/ephy-node.h:
        * src/bookmarks/ephy-bookmark-action.c:
        (bookmarks_child_changed_cb):
        * src/bookmarks/ephy-bookmark-properties.c:
        (toolbar_checkbox_changed_cb), (build_ui):
        * src/bookmarks/ephy-bookmarks-editor.c:
        (cmd_show_in_bookmarks_bar), (ephy_bookmarks_editor_update_menu):
        * src/bookmarks/ephy-bookmarks-menu.c: (create_menu),
        (topic_child_changed_cb):
        * src/bookmarks/ephy-new-bookmark.c:
        * src/bookmarks/ephy-new-bookmark.h:
        * src/bookmarks/ephy-topic-action.c:
        (ephy_topic_action_set_topic_id), (topic_child_changed_cb):

        Use guint consistently for node ids

+++ lib/ephy-node.h
@@ -51,12 +51,12 @@ typedef void (*EphyNodeCallback) (EphyNode *node, ...);
 EphyNode   *ephy_node_new                   (EphyNodeDb *db);
 
 EphyNode   *ephy_node_new_with_id           (EphyNodeDb *db,
-					     gulong reserved_id);
+					     guint reserved_id);
 
 EphyNodeDb *ephy_node_get_db		    (EphyNode *node);
 
 /* unique node ID */
-long        ephy_node_get_id                (EphyNode *node);
+guint       ephy_node_get_id                (EphyNode *node);
 
 /* refcounting */
 void        ephy_node_ref                   (EphyNode *node);
